requirements.txt
requests
supabase
python-dotenv
pandas
aiohttp
tqdm

